+++ project/data/skybox.vert
@@ -1,12 +1,12 @@
-uniform lowp mat4 ViewMatrix;
-uniform lowp mat4 ModelMatrix;
-uniform lowp mat4 ProjectionMatrix;
+uniform highp mat4 ViewMatrix;
+uniform highp mat4 ModelMatrix;
+uniform highp mat4 ProjectionMatrix;
 
 attribute vec4 Position;
 attribute vec4 TexCoord;
 
-varying lowp vec4 texCoordVarying;
-varying lowp vec4 posVarying;
+varying highp vec4 texCoordVarying;
+varying highp vec4 posVarying;
 
 void main()
 {
